remove the old similarity test
