+++ test/unit_test/test_similarity.py
@@ -1,65 +1,144 @@
 import pandas as pd
-from pandas.util.testing import assert_frame_equal
 
 from lexos.helpers.error_messages import NON_NEGATIVE_INDEX_MESSAGE
-from lexos.processors.analyze.similarity import similarity_maker
+from lexos.models.similarity_model import SimilarityModel, SimilarityTestOption
+from lexos.receivers.similarity_receiver import SimilarityOption
 
 
-class TestSimilarity:
-    def test_with_similarity_equals_one(self):
-        dtm_data_frame = pd.DataFrame([[0.0, 0.0, 9.0, 9.0, 0.0, 0.0, 5.0, 4.0,
-                                        0.0, 9.0, 0.0, 0.0, 0.0, 0.0, 0.0, 9.0,
-                                        0.0, 0.0, 0.0],
-                                       [0.0, 0.0, 9.0, 9.0, 0.0, 0.0, 0.0, 4.0,
-                                        5.0, 9.0, 0.0, 0.0, 0.0, 0.0, 0.0, 9.0,
-                                        0.0, 0.0, 0.0],
-                                       [5.0, 10.0, 0.0, 0.0, 10.0, 5.0, 0.0,
-                                        0.0, 0.0, 0.0, 10.0, 5.0, 5.0, 5.0,
-                                        5.0, 0.0, 5.0, 5.0, 5.0]],
-                                      index=['catBobcat', 'catCaterpillar',
-                                             'wake'])
-        comp_file_index = 2
-        assert_frame_equal(similarity_maker(dtm_data_frame, comp_file_index),
-                           pd.DataFrame([[1.0], [1.0]],
-                                        index=['catBobcat', 'catCaterpillar'],
-                                        columns=["Cosine similarity"]))
+# --------------------- test with similarity equals one --------------------
+# noinspection PyProtectedMember
+def test_with_similarity_equal_one():
+    test_dtm = pd.DataFrame([[0.0, 0.0, 9.0, 9.0, 0.0, 0.0, 5.0, 4.0, 0.0, 9.0,
+                              0.0, 0.0, 0.0, 0.0, 0.0, 9.0, 0.0, 0.0, 0.0],
+                             [0.0, 0.0, 9.0, 9.0, 0.0, 0.0, 0.0, 4.0, 5.0, 9.0,
+                              0.0, 0.0, 0.0, 0.0, 0.0, 9.0, 0.0, 0.0, 0.0],
+                             [5.0, 10.0, 0.0, 0.0, 10.0, 5.0, 0.0, 0.0, 0.0,
+                              0.0,
+                              10.0, 5.0, 5.0, 5.0, 5.0, 0.0, 5.0, 5.0, 5.0]],
+                            index=[0, 1, 2])
+    test_front_end_option = SimilarityOption(comp_file_id=2)
+    test_id_table = {0: "F1.txt", 1: "F2.txt", 2: "F3.txt"}
+    similarity_model = SimilarityModel(
+        test_options=SimilarityTestOption(
+            doc_term_matrix=test_dtm,
+            front_end_option=test_front_end_option,
+            id_temp_label_map=test_id_table
+        )
+    )
+    pd.testing.assert_series_equal(
+        similarity_model._gen_exact_similarity(),
+        pd.Series([1., 1.], index=["F1.txt", "F2.txt"])
+    )
+# --------------------------------------------------------------------------
 
-    def test_with_all_same_content_file(self):
-        dtm_data_frame = pd.DataFrame([[9.0, 9.0, 5.0, 4.0, 0.0, 9.0, 9.0],
-                                       [9.0, 9.0, 5.0, 4.0, 0.0, 9.0, 9.0],
-                                       [9.0, 9.0, 5.0, 4.0, 0.0, 9.0, 9.0]],
-                                      index=['file_1', 'file_2', 'file_3'])
-        comp_file_index = 1
-        assert_frame_equal(similarity_maker(dtm_data_frame, comp_file_index),
-                           pd.DataFrame([[0.0], [0.0]],
-                                        index=['file_1', 'file_3'],
-                                        columns=["Cosine similarity"]))
 
-    def test_with_two_dimension(self):
-        dtm_data_frame = pd.DataFrame([[0.0, 1.0], [1.0, 2.0], [2.0, 1.0]],
-                                      index=['file_1', 'file_2', 'file_3'])
-        comp_file_index = 0
-        assert_frame_equal(similarity_maker(dtm_data_frame, comp_file_index),
-                           pd.DataFrame([[0.1056], [0.5528]],
-                                        index=['file_2', 'file_3'],
-                                        columns=["Cosine similarity"]))
+# --------------------- test with all same content -------------------------
+# noinspection PyProtectedMember
+def test_with_all_same_content_file():
+    test_dtm = pd.DataFrame([[9.0, 9.0, 5.0, 4.0, 0.0, 9.0, 9.0],
+                             [9.0, 9.0, 5.0, 4.0, 0.0, 9.0, 9.0],
+                             [9.0, 9.0, 5.0, 4.0, 0.0, 9.0, 9.0]],
+                            index=[0, 1, 2])
+    test_front_end_option = SimilarityOption(comp_file_id=1)
+    test_id_table = {0: "F1.txt", 1: "F2.txt", 2: "F3.txt"}
+    similarity_model = SimilarityModel(
+        test_options=SimilarityTestOption(
+            doc_term_matrix=test_dtm,
+            front_end_option=test_front_end_option,
+            id_temp_label_map=test_id_table
+        )
+    )
+    pd.testing.assert_series_equal(
+        similarity_model._gen_exact_similarity(),
+        pd.Series([0., 0.], index=["F1.txt", "F3.txt"])
+    )
+# --------------------------------------------------------------------------
 
-    def test_with_three_dimension(self):
-        dtm_data_frame = pd.DataFrame([[1.0, 1.0, 1.0],
-                                       [1.0, 0.0, 0.0],
-                                       [0.0, 2.0, 1.0]],
-                                      index=['file_1', 'file_2', 'file_3'])
-        comp_file_index = 1
-        assert_frame_equal(similarity_maker(dtm_data_frame, comp_file_index),
-                           pd.DataFrame([[0.4226], [1.0]],
-                                        index=['file_1', 'file_3'],
-                                        columns=["Cosine similarity"]))
 
-    def test_similarity_maker_non_neg_index_precondition(self):
-        try:
-            dtm_data_frame = pd.DataFrame([[1.0], [1.0]],
-                                          index=['test_1', 'test_2'])
-            _ = similarity_maker(dtm_data_frame, comp_file_index=-1)
-            raise AssertionError("negative index error did not raise.")
-        except AssertionError as error:
-            assert str(error) == NON_NEGATIVE_INDEX_MESSAGE
+# --------------------- test with with two dimension -----------------------
+# noinspection PyProtectedMember
+def test_with_two_dimension():
+    test_dtm = pd.DataFrame([[0.0, 1.0], [1.0, 2.0], [2.0, 1.0]],
+                            index=[0, 1, 2])
+    test_front_end_option = SimilarityOption(comp_file_id=0)
+    test_id_table = {0: "F1.txt", 1: "F2.txt", 2: "F3.txt"}
+    similarity_model = SimilarityModel(
+        test_options=SimilarityTestOption(
+            doc_term_matrix=test_dtm,
+            front_end_option=test_front_end_option,
+            id_temp_label_map=test_id_table
+        )
+    )
+    # assertion
+    pd.testing.assert_series_equal(
+        similarity_model._gen_exact_similarity(),
+        pd.Series([.105572809, .5527864045],
+                  index=["F2.txt", "F3.txt"],
+                  name="cos_similarity")
+    )
+# --------------------------------------------------------------------------
+
+
+# --------------------- test with with three dimension ---------------------
+# noinspection PyProtectedMember
+def test_with_three_dimension():
+    test_dtm = pd.DataFrame([[1.0, 1.0, 1.0], [1.0, 0.0, 0.0],
+                             [0.0, 2.0, 1.0]],
+                            index=[0, 1, 2])
+    test_front_end_option = SimilarityOption(comp_file_id=1)
+    test_id_table = {0: "F1.txt", 1: "F2.txt", 2: "F3.txt"}
+    similarity_model = SimilarityModel(
+        test_options=SimilarityTestOption(
+            doc_term_matrix=test_dtm,
+            front_end_option=test_front_end_option,
+            id_temp_label_map=test_id_table
+        )
+    )
+    # assertion
+    pd.testing.assert_series_equal(
+        similarity_model._gen_exact_similarity(),
+        pd.Series([.42264973081, 1.],
+                  index=["F1.txt", "F3.txt"],
+                  name="cos_similarity")
+    )
+# --------------------------------------------------------------------------
+
+
+# --------------------- test with with special case ------------------------
+# noinspection PyProtectedMember
+def test_with_special_case_one():
+    try:
+        test_dtm = pd.DataFrame([[1.0], [1.0]], index=[0, 1])
+        test_front_end_option = SimilarityOption(comp_file_id=-1)
+        test_id_table = {0: "F1.txt", 1: "F2.txt"}
+        similarity_model = SimilarityModel(
+            test_options=SimilarityTestOption(
+                doc_term_matrix=test_dtm,
+                front_end_option=test_front_end_option,
+                id_temp_label_map=test_id_table
+            )
+        )
+        _ = similarity_model._gen_exact_similarity()
+        raise AssertionError("negative index error did not raise.")
+    except AssertionError as error:
+        assert str(error) == NON_NEGATIVE_INDEX_MESSAGE
+
+
+# noinspection PyProtectedMember
+def test_with_special_case_two():
+    try:
+        test_dtm = pd.DataFrame([[1.0], [1.0]], index=[0, 1])
+        test_front_end_option = SimilarityOption(comp_file_id=-2)
+        test_id_table = {0: "F1.txt", 1: "F2.txt"}
+        similarity_model = SimilarityModel(
+            test_options=SimilarityTestOption(
+                doc_term_matrix=test_dtm,
+                front_end_option=test_front_end_option,
+                id_temp_label_map=test_id_table
+            )
+        )
+        _ = similarity_model._gen_exact_similarity()
+        raise AssertionError("negative index error did not raise.")
+    except AssertionError as error:
+        assert str(error) == NON_NEGATIVE_INDEX_MESSAGE
+# --------------------------------------------------------------------------
